requirements.txt
asgiref==3.4.1
Django==3.2.8
mysql-connector-python==8.0.27
mysqlclient==2.0.3
protobuf==3.19.1
pytz==2021.3
sqlparse==0.4.2
